Fixed iteration (pointer arith / size_t)

+++ ft_strmap.c
@@ -6,7 +6,7 @@
 /*   By: adavis <adavis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/10 17:17:23 by adavis            #+#    #+#             */
-/*   Updated: 2019/04/24 16:30:50 by adavis           ###   ########.fr       */
+/*   Updated: 2019/04/25 16:34:14 by adavis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 char	*ft_strmap(char const *s, char (*f)(char))
 {
 	char	*sn;
-	int		i;
+	size_t	i;
 
 	if (!s || !f)
 		return (0);
