+++ ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: adavis <adavis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/28 15:37:43 by adavis            #+#    #+#             */
-/*   Updated: 2019/04/09 16:43:37 by adavis           ###   ########.fr       */
+/*   Updated: 2019/04/25 16:35:04 by adavis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dst, const char *src, size_t len)
 {
-	unsigned long	i;
+	size_t	i;
 
 	i = 0;
 	while (i < len && src[i] != '\0')
