+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: adavis <adavis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/23 18:15:19 by adavis            #+#    #+#             */
-/*   Updated: 2019/04/22 15:36:35 by adavis           ###   ########.fr       */
+/*   Updated: 2019/04/25 16:38:26 by adavis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,7 +79,7 @@ int					ft_isspace(char c);
 void				ft_strrev(char *str);
 char				*ft_itoa(int n);
 int					ft_strwrdcnt(char *str, char sep);
-int					ft_strwrdlen(char *str, char sep);
+size_t				ft_strwrdlen(char *str, char sep);
 char				**ft_strsplit(char const *s, char c);
 size_t				ft_strlcat(char *dst, const char *src, size_t size);
 t_list				*ft_lstnew(void const *content, size_t content_size);
